gzip: fix memory allocs (buffers not returned to pool) (#28)
* gzip: fix allocation leak during Write

Fix allocation leaks during gzip writes. Due to incorrect use of the
dstPool (unmatched Get and Put), large amounts of memory were
temporarily allocated during writes and not put back into the pool.

This also removes some special handling code in compressCurrent that
would recursively call itself for too large input buffers. This
condition can never occur, because Write ensures that blocks are capped
and there is no other public interface that extends currentBuffer.
The recursive call that slices the buffer would have made returning byte
slices to the Pool dangerous, as we could have been returning the same
underlying buffer multiple times.

This also adds a test to check allocations per Write to prevent
regressions. There is further room for improvement, but this was by far
the biggest leak.

* gzip tests: fix allocation test

The allocation test was incorrectly using Alloc (the current number of
allocated bytes) instead of TotalAlloc (the cummulative number of
bytes), which would break the test if garbage collection happened during
the run.

This also increases the allocated bytes check to allow for occasional
other background allocations. I once saw 3kB instead of the usual <700B
allocations. This is still sufficiently low to detect pool usage issues.

* gzip: make clear panic is due to concurrent Write race

In the panics, make clear that these can only happen due to concurrent
Write call races, which are unsafe. The alternative would be to silently
corrupt data.

Move the panic to the beginning of compressCurrent to not deadlock tests
after a panic.

* gzip: fix race in prevTail

Use a pool buffer for prevTail instead of slicing the current buffer,
which may have been returned to the pool by the time the tail is
accessed.

* gzip tests: do not run alloc test with -race

Skip the allocation tests when running with the race detector, because
the race detector allocates a lot of memory.

Rename the unreliable tests file.

* gzip tests: private allocBytesPerRun